requirements.txt
pyyaml>=6.0
httpx>=0.27.0
python-dotenv>=1.0.0
openai>=1.0.0
langgraph>=1.1.10
aiohttp>=3.9.0
Pillow>=10.0.0